requirements.txt
aniso8601==7.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.0
Flask-RESTful==0.3.7
get==2019.4.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
post==2019.4.13
public==2019.4.13
pytz==2019.1
PyYAML==5.1.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
